+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to BrightCast will be documented in this file.
 - Ability to download songs.
 - Handling vk error 5 (User authorization failed) and showing instruction how to fix it
 - App shortcuts (ctrl + arrow left - skip prev, ctrl + arrow right - skip next, ctrl + space - play/pause, ctrl + arrow up - increase volume, ctrl + arrow down - decrease volume)
+- Showing error message on mobile when connection with desktop lost
 
 ### Removed
 - Cleaning users' broadcast on app close.